requirements.txt
Flask==1.1.1
requests==2.23.0
python-dotenv==0.11.0
pusher==2.1.4
RPi.GPIO==0.7.0